Status of commands for Allan to see. Works but is not pretty.

+++ examples/vslice/src/main/scala/csw/examples/vslice/assembly/TromboneAssembly.scala
@@ -184,7 +184,7 @@ object TromboneAssembly {
   // Move submit command
   val movePrefix = s"$componentPrefix.move"
   val moveCK:ConfigKey = movePrefix
-  def moveSC(position: Int):SetupConfig = SetupConfig(moveCK).add(stagePositionKey -> position withUnits stagePositionUnits)
+  def moveSC(position: Double):SetupConfig = SetupConfig(moveCK).add(stagePositionKey -> position withUnits stagePositionUnits)
 
     // Position submit command
   val positionPrefix = s"$componentPrefix.position"
